fix(domain/Coin): use float not int for price

+++ internal/core/domain/domain.go
@@ -5,7 +5,7 @@ import "time"
 type Coin struct {
   ID string
   Symbol string
-  Price int16
+  Price float32
   MarketCap int64
   UpdatedAt time.Time
 }
